+++ web/documentserver-example/java/src/main/java/entities/FileType.java
@@ -33,5 +33,8 @@ public enum FileType {
     SLIDE,
     @JsonProperty("pdf")
     @SerializedName("pdf")
-    PDF
+    PDF,
+    @JsonProperty("diagram")
+    @SerializedName("diagram")
+    DIAGRAM
 }